+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## [main] 0.1.10
 - 修复 issue #48: 修复错误的引导路径。
-- 修复非中文语言下，初始化引导界面不跟随主界面的问题。
-- 修复部分模型的 usage 无法正常显示的问题。
-- 实现 openmcp 工具测试的并行实现和暂停功能。
+- 支持 kimi 的 usage 计数 + 支持 kimi 的 system prompt。
+- 实现 openmcp 工具测试的并行执行和暂停功能。 https://picx.zhimg.com/80/v2-4e09958d91dcf561c578294d8b6f3349_1440w.png
+- 修正 API 测速中算法，剥离为 tps + 排队时间两部分。 https://picx.zhimg.com/80/v2-1cc3044a3ec3d5d21cb265dd67518ca0_1440w.png
+- 大模型 api 测速目前可以自定义 prompt 了。 https://picx.zhimg.com/80/v2-ff70af72254b82c11a941fe9cc29eeb8_1440w.png
+- 实现 issue#49，工具模块，调试希望支持markdown渲染回显。 https://picx.zhimg.com/80/v2-5d708ccab00f33fdf63a656a0066bf23_1440w.png
+- 实现 issue#54，右击服务器列表名，可以重命名服务器。 https://picx.zhimg.com/80/v2-87c2a29abdd2dd56a4d18cc4a8b946ff_1440w.png
+- 修复 resources 和 prompts 有关热更新的一些问题。
+- 更新 vscode 创建的标签栏的标题。 https://picx.zhimg.com/80/v2-4d40c20f3eaa032573e4de58298c859f_1440w.png
 
 ## [main] 0.1.9
 - 增加 mook 功能：可以利用随机种子或者AI生成来自动化填充测试 tool 的表单数据